Version 1.0
Bunch of items for NPC'S I have in my campaign. Also none of this stuff is balance tested, I just kinda asked AI if it though it was balanced based on a similar class/subclass and listened to what adjustments it told me to do.

1 custom class w/ 5 subclasses
- Wyldheart (BG3 Mod that I redid)
-- Master of Hounds
-- Greenspeaker
-- Kaos Warden
-- Primal Lord
-- Ravensong

The companions show up as cantrips, because I don't feel like messing with companions yet. I don't know enough.

3 Barb Subclasses
- Blackblood (Crona from Soul Eater) (Crimson Original)
- Lycanthrope (Crimson Original)
- Vituoso (Crimson Original)

1 Cleric Subclass
- Blood (Crimson Original)

2 Fighter Subclasses
- Flowmtion (based on Sora's Flowmotion)(Crimson Original)
-Limitbreaker (based on a mix of abilities from Cloud and Noctis from the FF series)(Crimson Original)

1 Ranger subclass
- Death Whistler (Based on Yondu)(Crimson Original)